+++ src/backend/linux_raw/fs/mod.rs
@@ -1,4 +1,3 @@
-#[cfg(any(feature = "fs", feature = "procfs"))]
 pub(crate) mod dir;
 pub(crate) mod makedev;
 pub(crate) mod syscalls;
